src/requirements.txt
BeautifulSoup4
requests
names
python-dateutil
numpy
# nltk